# docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "Contacts app",
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "description": "This is a documentation of contacts app"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:3000"
    },
    {
      "url": "https://nodejs-hw-mongodb-5f0u.onrender.com"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
    "/contacts": {
      "get": {
        "tags": [
          "Contacts"
        ],
        "summary": "Get a list of contacts",
        "operationId": "getContacts",
        "description": "Retrieve a list of contacts with optional filtering.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/page"
          },
          {
            "$ref": "#/components/parameters/perPage"
          },
          {
            "$ref": "#/components/parameters/sortBy"
          },
          {
            "$ref": "#/components/parameters/sortOrder"
          },
          {
            "$ref": "#/components/parameters/type"
          },
          {
            "$ref": "#/components/parameters/isFavourite"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/200"
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      },
      "post": {
        "tags": [
          "Contacts"
        ],
        "summary": "Create a new contact",
        "operationId": "createContact",
        "description": "This endpoint allows you to create a new contact.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "name",
                  "email",
                  "phoneNumber"
                ],
                "properties": {
                  "name": {
                    "type": "string",
                    "description": "Name of the contact",
                    "example": "John Doe"
                  },
                  "email": {
                    "type": "string",
                    "description": "Email address of the contact",
                    "example": "john.doe@example.com"
                  },
                  "phoneNumber": {
                    "type": "string",
                    "description": "Phone number of the contact",
                    "example": "+123456789"
                  },
                  "isFavourite": {
                    "type": "boolean",
                    "description": "Indicates if the contact is marked as favourite",
                    "default": false
                  },
                  "contactType": {
                    "type": "string",
                    "enum": [
                      "work",
                      "home",
                      "personal"
                    ],
                    "description": "Type of contact",
                    "example": "work"
                  },
                  "photo": {
                    "type": "string",
                    "format": "binary",
                    "description": "Contact's photo (optional)"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "$ref": "#/components/responses/201"
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    },
    "/contacts/{id}": {
      "get": {
        "tags": [
          "Contacts"
        ],
        "operationId": "getContactById",
        "summary": "Retrieve a contact's full information by its unique identifier.",
        "description": "This endpoint allows you to get a contact by its unique ID.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID of the contact to retrieve",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved the contact",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Contact"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "404": {
            "$ref": "#/components/responses/404"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      },
      "patch": {
        "tags": [
          "Contacts"
        ],
        "operationId": "updateContact",
        "summary": "Update contact information by ID",
        "description": "This endpoint allows you to update specific contact details by its unique ID.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID of the contact to update",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "John Doe"
                  },
                  "email": {
                    "type": "string",
                    "format": "email",
                    "example": "johndoe@example.com"
                  },
                  "phoneNumber": {
                    "type": "string",
                    "example": "+1234567890"
                  },
                  "photo": {
                    "type": "string",
                    "format": "binary",
                    "description": "Contact's updated photo (optional)"
                  },
                  "contactType": {
                    "type": "string",
                    "enum": [
                      "work",
                      "home",
                      "personal"
                    ],
                    "example": "personal"
                  },
                  "isFavourite": {
                    "type": "boolean",
                    "description": "Indicates if the contact is marked as favourite",
                    "example": true
                  }
                }
              },
              "example": {
                "email": "tetetetete@gmail.com",
                "contactType": "home",
                "phoneNumber": "+123456789",
                "name": "Jane Doe",
                "photo": "file data",
                "isFavourite": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successfully updated contact details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Contact"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "404": {
            "$ref": "#/components/responses/404"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      },
      "delete": {
        "tags": [
          "Contacts"
        ],
        "operationId": "deleteContactById",
        "summary": "Delete a contact by ID",
        "description": "Delete a contact by its unique ID.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID of the contact to delete",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "$ref": "#/components/responses/204"
          },
          "404": {
            "$ref": "#/components/responses/404"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "responses": {
      "200": {
        "description": "Request was successful",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "integer",
                  "example": 200
                },
                "message": {
                  "type": "string",
                  "example": "Successfully found contacts!"
                },
                "data": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "_id": {
                            "type": "string",
                            "description": "The unique identifier of the contact",
                            "example": "67f25d39e1935b9687129f29"
                          },
                          "name": {
                            "type": "string",
                            "example": "John Doe"
                          },
                          "phoneNumber": {
                            "type": "string",
                            "example": "+123456789"
                          },
                          "email": {
                            "type": "string",
                            "example": "john.doe@example.com"
                          },
                          "isFavourite": {
                            "type": "boolean",
                            "example": false
                          },
                          "contactType": {
                            "type": "string",
                            "example": "work"
                          },
                          "userId": {
                            "type": "string",
                            "example": "67f22a41875fd7d41ec03fd3"
                          },
                          "photo": {
                            "type": "string",
                            "format": "binary",
                            "example": null
                          },
                          "createdAt": {
                            "type": "string",
                            "format": "date-time",
                            "example": "2025-04-06T10:53:45.112Z"
                          },
                          "updatedAt": {
                            "type": "string",
                            "format": "date-time",
                            "example": "2025-04-06T10:53:45.112Z"
                          }
                        }
                      }
                    },
                    "page": {
                      "type": "integer",
                      "example": 1
                    },
                    "perPage": {
                      "type": "integer",
                      "example": 10
                    },
                    "totalItems": {
                      "type": "integer",
                      "example": 4
                    },
                    "totalPages": {
                      "type": "integer",
                      "example": 1
                    },
                    "hasPreviousPage": {
                      "type": "boolean",
                      "example": false
                    },
                    "hasNextPage": {
                      "type": "boolean",
                      "example": false
                    }
                  }
                }
              }
            }
          }
        }
      },
      "201": {
        "description": "Resource created successfully",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "integer",
                  "example": 201
                },
                "message": {
                  "type": "string",
                  "example": "Successfully created a contact!"
                },
                "data": {
                  "type": "object",
                  "properties": {
                    "_id": {
                      "type": "string",
                      "example": "67f3cabf966746f5cc383df4"
                    },
                    "name": {
                      "type": "string",
                      "example": "John Doe"
                    },
                    "phoneNumber": {
                      "type": "string",
                      "example": "+1234567890"
                    },
                    "email": {
                      "type": "string",
                      "example": "john.doWWW@example.com"
                    },
                    "isFavourite": {
                      "type": "boolean",
                      "example": true
                    },
                    "contactType": {
                      "type": "string",
                      "enum": [
                        "work",
                        "home",
                        "personal"
                      ],
                      "example": "personal"
                    },
                    "userId": {
                      "type": "string",
                      "example": "67f22a41875fd7d41ec03fd3"
                    },
                    "photo": {
                      "type": "string",
                      "format": "uri",
                      "example": "https://res.cloudinary.com/dl22iuo55/image/upload/v1744030398/jy3u6b7thbwnfgt2n6jm.webp"
                    },
                    "createdAt": {
                      "type": "string",
                      "format": "date-time",
                      "example": "2025-04-07T12:53:19.153Z"
                    },
                    "updatedAt": {
                      "type": "string",
                      "format": "date-time",
                      "example": "2025-04-07T12:53:19.153Z"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "204": {
        "description": "Resource deleted successfully. No content returned."
      },
      "400": {
        "description": "Invalid request or bad input",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponseSchema"
            },
            "example": {
              "status": 400,
              "message": "Something went wrong",
              "data": "\"email\" is required, \"emai\" is not allowed"
            }
          }
        }
      },
      "401": {
        "description": "Unauthorized access",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponseSchema"
            },
            "example": {
              "status": 401,
              "message": "Not authenticated"
            }
          }
        }
      },
...
      "404": {
        "description": "Resource not found",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string"
                }
              }
            },
            "example": {
              "message": "Not found"
            }
          }
        }
      },
...
      "500": {
        "description": "Internal server error. Something went wrong on the server.",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "integer",
                  "example": 500
                },
                "message": {
                  "type": "string",
                  "example": "Something went wrong"
                },
                "data": {
                  "type": "string",
                  "example": "Error details"
                }
              }
            },
            "example": {
              "status": 500,
              "message": "Something went wrong",
              "data": "Error details"
            }
          }
        }
      },
...
    },
    "parameters": {
      "page": {
        "name": "page",
        "in": "query",
        "description": "Page number",
        "required": false,
        "schema": {
          "type": "integer",
          "minimum": 1,
          "default": 1
        }
      },
      "perPage": {
        "name": "perPage",
        "in": "query",
        "description": "Number of items per page",
        "required": false,
        "schema": {
          "type": "integer",
          "minimum": 1,
          "default": 10
        }
      },
      "sortBy": {
        "name": "sortBy",
        "in": "query",
        "description": "Sort field",
        "required": false,
        "schema": {
          "type": "string",
          "enum": [
            "name"
          ],
          "default": "name"
        }
      },
      "sortOrder": {
        "name": "sortOrder",
        "in": "query",
        "description": "Sort order",
        "required": false,
        "schema": {
          "type": "string",
          "enum": [
            "asc",
            "desc"
          ],
          "default": "asc"
        }
      },
      "type": {
        "name": "type",
        "in": "query",
        "description": "Contact type filter",
        "required": false,
        "schema": {
          "type": "string",
          "enum": [
            "work",
            "home",
            "personal"
          ]
        }
      },
      "isFavourite": {
        "name": "isFavourite",
        "in": "query",
        "description": "Filter by favourite status",
        "required": false,
        "schema": {
          "type": "boolean"
        }
      }
    },
    "schemas": {
      "ErrorResponseSchema": {
        "type": "object",
        "properties": {
          "status": {
            "type": "integer",
            "example": 401
          },
          "message": {
            "type": "string",
            "example": "Unauthorized"
          },
          "data": {
            "type": "string",
            "example": "\"email\" is required, \"emai\" is not allowed"
          }
        },
        "required": [
          "status",
          "message"
        ]
      },
      "Contact": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "example": "67f25d39e1935b9687129f29"
          },
          "name": {
            "type": "string",
            "example": "John Doe"
          },
          "email": {
            "type": "string",
            "example": "john.doe@example.com"
          },
          "phoneNumber": {
            "type": "string",
            "example": "+123456789"
          },
          "isFavourite": {
            "type": "boolean",
            "example": false
          },
          "contactType": {
            "type": "string",
            "example": "work"
          },
          "userId": {
            "type": "string",
            "example": "67f22a41875fd7d41ec03fd3"
          },
          "photo": {
            "type": "string",
            "format": "binary",
            "example": null
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "example": "2025-04-06T10:53:45.112Z"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "example": "2025-04-06T10:53:45.112Z"
          }
        },
        "required": [
          "_id",
          "name",
          "email",
          "phoneNumber",
          "isFavourite",
          "contactType",
          "userId",
          "createdAt",
          "updatedAt"
        ]
      },
...
    }
  }
}